Switch thread_pool on ruby platform internally
